(4) Segmentation/requirements.txt
Pillow>=8.0.0
ttkthemes>=3.2.0